feat: 增强异常处理信息
使用BussinessException提供更详细的错误信息

+++ src/main/java/top/tankimzeg/editorial_system/service/UserService.java
@@ -1,8 +1,10 @@
 package top.tankimzeg.editorial_system.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import top.tankimzeg.editorial_system.entity.User;
+import top.tankimzeg.editorial_system.exception.BusinessException;
 import top.tankimzeg.editorial_system.repository.UserRepo;
 
 import java.time.LocalDateTime;
@@ -32,6 +34,9 @@ public class UserService {
     }
 
     public void deleteUserById(Long id) {
+        if (!userRepo.existsById(id)) {
+            throw new BusinessException(HttpStatus.NOT_FOUND, "用户不存在");
+        }
         userRepo.deleteById(id);
     }
